[Gate Fix] Fix AttributeError in ServerActionsRbacTest
This PS fixes an AttributeError being raised in ServerActionsRbacTest
[0] due to a recent Tempest change [1]. Besides, the volume
creation was being done manually in Patrole, when the volume
should have been created using Tempest's base helper for creating
a volume.

This same issue applies with `test_create_image_volume_backed` in
the same class [2] due to a recent Tempest change [3], and besides
that, the previous way of spinning up a volume-backed server did
not leverage Tempest's helpers, which this PS corrects.

[0] http://logs.openstack.org/06/466806/8/check/gate-tempest-dsvm-patrole-admin-ubuntu-xenial/c186718/console.html#_2017-09-22_01_07_10_242418
[1] https://review.openstack.org/#/c/498888/
[2] http://logs.openstack.org/31/506531/3/check/gate-tempest-dsvm-patrole-member-ubuntu-xenial/01bfbc6/console.html#_2017-09-22_07_08_39_483859
[3] https://review.openstack.org/#/c/498886/

Change-Id: I6f93dc6961844e2154d348790cad3fde77b9d218