Clean up exception message raised by policy authority module
This PS cleans up an exception message raised by policy authority
module because it is innacurate. When a policy isn't found,
the current message says something like:

Policy action: group:reset_group_snapshot not found in policy file:
None. [0]

However, this isn't very helpful as in this case no custom policy
file is used and besides that, the policy in code is used as a source
for finding the policy. So the message has been changed to (e.g.):

Policy action "group:reset_group_snapshot" not found in policy file:
None or among registered policy in code defaults for service.

[0] http://logs.openstack.org/12/536612/2/check/patrole-admin/7367a52/job-output.txt.gz#_2018-01-23_00_15_27_996109

Change-Id: I754a03d57ae404877ad7521e25bf49a5cc7357b9