Remove admin_only kwarg from rbac_rule_validation decorator
The ``admin_only`` kwarg has been removed from ``rbac_rule_validation``
decorator because it is no longer used by any tests. Besides that,
it should not be used because Patrole is dedicated to RBAC testing and
an admin-only check is not RBAC because it does not use ``oslo.policy``
library.

Change-Id: Iee4881a8756cf2f2a0a4520dd450776a4f1e873c