Fix test_limits_rbac not checking for expected response
This is to fix test_show_limits test under test_limits_rbac
in volume module which is currently expecting a fatal exception
to be raised by the Cinder server following failed authorization.
However, this assumption is false because this endpoint uses
non-fatal authorization [0] to inject additional attributes into
the response body following successfuly authorization.

Thus, it is necessary Patrole side to verify that the expected
attributes are present in the response body, as expecting a 403
is a false assumption.

[0] https://github.com/openstack/cinder/blob/67177026ab510b09644fd4b8b301594aa5482cef/cinder/api/contrib/used_limits.py#L30

Change-Id: I3aae5d8c6c947cf5119f1cdc081dae8179421564