Remove already-deprecated strict_policy_check option
The option ``[patrole].strict_policy_check`` was deprecated
during the last release cycle (Queens). This was because the
option could be set to False in order to skip tests which
might result in false positives.

This PS, then, removes strict_policy_check references in the code,
updates documentation, and adds a releasenote.

Change-Id: I7f7eda39c0472bd3d70892c801fc4d14db0c0426