[Fix Gate] Fix network segments policy test raising 503
This is to fix the gate by resolving an issue with a Neutron
segments policy test class that hardcodes the network segment
to use gre network type, which may not be supported by the
underlying network, resulting in a 503 exception being raised.
Example stacktrace: [0] and root cause: [1].

Also renames the class and file to convey that actual policy
being tested.

[0] http://logs.openstack.org/20/547820/7/check/patrole-admin/5f4e495/job-output.txt.gz#_2018-03-12_20_50_15_600175
[1] http://logs.openstack.org/20/547820/7/check/patrole-admin/5f4e495/controller/logs/screen-neutron-api.txt.gz#_Mar_12_20_50_15_698713

Change-Id: I6e3daef766f0ed5fabcc125f39471d089c9d7218