Increase test stability of backup RBAC tests
This is to increase test stability of backup RBAC tests which
fail not only due to the rather error-prone nature of backing
up volumes, but also because of data race conditions and
missing clean ups.

This adds some waiters following certain backup actions. This
also adds a clean up of the volume that is created when restoring
a backup [0].

Additional waiters are also added for waiting for volumes to
become available following backup (because they briefly
go into backing-up status) [1].

[0] https://developer.openstack.org/api-ref/block-storage/v3/index.html#restore-a-backup
[1] https://review.openstack.org/#/c/569917/8

Change-Id: Ifce15f69fcca38005b40785c42231ca164917345