Fix: Set max_microversion to 2.38 for image metadata compute tests
This is to set the max_microversion to 2.38 for image metadata
compute tests which is a deprecate Glance proxy API deprecated
in 2.39. The code incorrectly includes it under the max_microversion
for the image proxy API whose max_microversion is 2.35. However,
2.38 is the correct microversion for image metadata tests,
not 2.35, which is for image tests.

For details, see: Ic6cc22f048c2ccd427dfe6e6aea0273128d96592

Partially Implements bp: clear-deprecated-api
Change-Id: Ia97f55b2004e5fa8abbab0365b6482cf42bda738