Rephrase Patrole skip message for better sorting
This is a trivial fix that is meant to address a quality of life
issue around sorting test results.

This addresses essentially the following concern:

We are starting to run Patrole and we are finding that it has a
lot different messages. We should reverse the message so they sort
better.

Change-Id: I3b5e2d9b1a9d1425982260e910a861d822625b06