Docs: Add requirements authority module to documentation
This patchset includes documentation on the rbac_authority
and the requirements_authority modules. In addition,
the documentation for the policy_authority module is
expanded. All 3 modules are explained together, explaining
that the rbac_authority module contains an abstract
class consumed by the classes in the other two modules.

The use cases for each validation approach is also included
in the documentation.

Finally, some documentation syntax issues are corrected.

Change-Id: I33bbe2da67683faafd0749b687b99237ac815009