Limit .zuul.yaml gate jobs to most important jobs only
This patchset limits the jobs enqueued in the gate pipeline in
Zuul to the 2 most important jobs only:

- patrole-admin
- patrole-member

Everything else will remain in the check pipeline. This means
that the chances of errant gate failures will be significantly
reduced. At the same time, test coverage will remain the same.
Now, however, all jobs will run once during the check pipeline
(when commits are first submitted) and, after cores approve
of a patchset, only the 2 jobs above (in addition to unit test
jobs, pep8, etc.) will run [0].

It is extremely unlikely that a patchset B that is merged between
patchset A's check and gate pipelines will cause a breaking change
to enter the source code -- and even so, this is an acceptable
margin of error when balanced against developer workflow being
slowed down in some cases severely due to many, many spurious
failures across volume and server APIs related to driver instability
(not RBAC), across multiple jobs enqueued in the gate pipeline.

[0] https://docs.openstack.org/infra/system-config/zuul.html

Change-Id: I57772e00d4e96a8ac841651b7f36379629940956