trivial: Drop empty services folder under patrole_tempest_plugin
This patchset drops the empty services folder under
patrole_tempest_plugin. While it is part of the cookie cutter
Tempest plugin setup, it is currently empty and serves no
purpose in Patrole. If in the future we decide to add service
clients there, we can reintroduce the folder. But there is
no purpose in keeping around an empty folder -- it just
increases clutter.

Change-Id: Ibf83a73f5dc2c38d737359221e9974a6562846a6