Improve exception that is raised following invalid service
This patchset changes the exception that is raised following
an invalid service being passed in for testing. Instead of
a NotFound getting raised (which doesn't make sense), the
original exception, RbacInvalidServiceException, is raised
instead. In addition "Exception" is appended to the original
exception name for greater clarity and consistency with other
exception names.

This patchset is similar to [0] which aims to simplify and
consolidate how service input is validated.

[0] https://review.openstack.org/#/c/577963/

Change-Id: Ib91a6581f89a528630d756840176b0e16663fa6f