Add documentation explaining why Patrole tests aren't in Tempest
There have been concerns raised as to why Patrole tests aren't
contained in Tempest (the concerns are not found in any written
discussion online). This documentation puts such concerns to
rest. It was agreed upon in the RBAC testing spec [0] that:
"rbac tests will live in separate tempest plugin." This
documentation formalizes that reasoning.

[0] comment in https://review.openstack.org/#/c/382672/

Change-Id: I31d956b42440a5448a5be0a7e2c5b3b7ddacfab5