Deprecate enable_rbac configuration option
The ``[patrole].enable_rbac`` option is deprecated and will be removed
during the "S" release. This is a legacy option that was meaningful
downstream when Patrole was a suite of tests inside Tempest itself. Now
that Patrole exists upstream as a Tempest plugin, it is paradoxical to
install the Patrole plugin yet have an option that allows all Patrole
tests to be skipped. This option is at odds with current Patrole
architecture.

To skip RBAC tests going forward, with Patrole Tempest plugin already
installed, use an appropriate regex.

Change-Id: I0cc39184c7891fb9133a45488a36ba220592915f