Remove misleading todo from config.py
This patch set removes a misleading todo from config.py which
suggests that it might be possible to leverage Keystone's
Policy API to derive policy information from remote hosts. But
this is not possible as the Policy API is deprecated [0].

If we want to get policy information from a remote host we will
need to pursue some SSH alternative instead, but this is not
planned anytime soon.

[0] https://docs.openstack.org/releasenotes/keystone/queens.html#deprecation-notes

Change-Id: I8a4da457c2d40ebddbfe5492dc1cbba87288c0da