Remove the usage of deprecated arg 'expected_error_code'
'expected_error_code' arg to rbac_rule_validation.action is
deprecated in favor of 'expected_error_codes'.

Many test use the old deprecated arg and that's why the warning. We
should cleanup the usage in patrole to avoid the confusion and make
all test to latest interface.

This commit migrate those usage to new arg and leave the deprecated
arg as it is which can be removed completely after deprecation phase.

Change-Id: Ie284ae74ce6480a59406aecfe341d579142bd625