Remove extra_attr kwarg from RbacMalformedResponse
This patch set removes extra_attr kwarg from RbacMalformedResponse
exception because it is unclear how unauthorized policy enforcement
could result in an extra attribute being injected into a response
body. For such an edge case to even be allowed, there would have
to be a documented example of this. Besides, no test case even
makes use of it, meaning that it is not something we should
even support.

Change-Id: I8f63720fe02d63360a7e40c836d55b56eb657156
Story: 2003843
Task: 26634