Deprecate use of v2 roles client in rbac_utils.py
This patch set deprecates use of the v2 roles client for
role overriding operations. This is because the Keystone v2 API
is deprecated and slated for removal, so Patrole's use of the
v2 Tempest roles client is also deprecated and will be
removed in a future release. Patrole will only support the v3 Tempest
roles client for role overriding operations in the future.

Change-Id: I9b201677cf8244b25f3cc8ea3b48a95b2b83b95e