Remove invalid exception RbacConflictingPolicies
This patch set drops RbacConflictingPolicies from Patrole because
its intention is 1) flawed and 2) the way it is being used is
invalid.

First, there is no documented case where conflicting policies
exists in OpenStack, so the intention behind the exception class
is flawed.

Second, the exception is only used in one test in
test_server_rbac. It catches a ServerFault (500) and re-raises
the exception is a pseudo-Forbidden exception. This is very bad
because if Nova raises a 500 on a policy exception then this is
a Nova bug. There is no valid use case where a ServerFault from
Nova should ever be treated as a successful test scenario.

Change-Id: I495b10ccf7b0df523b2551c49ea4db07e5bcaf79