fix: Rename test classes causing accidental skip
This patch set renames two classes that are neutron plugin
rbac tests but don't end in the appropriate suffix of
"PluginRbacTest" which is used by .zuul.yaml to correctly execute
the tests [0].

This means that, in effect, these tests aren't being run in any
jobs. This patch set resolves the issue.

In a follow up patch set a hacking rule will be added to enforce
correct naming for plugin rbac tests.

[0] https://github.com/openstack/patrole/blob/0c7f83ea1716462737a40cc32f27b95abf4557af/.zuul.yaml#L168

Change-Id: Ieac5ebf70eefcebf1aebc26da95d58fcd68a202e