refactor: Rename PluginRbacTest => ExtRbacTest
This patch set is a follow up on discussion in [0] which
concluded that Patrole should not test Neutron plugins.

The pertinent discussion from [0] is:

Patrole can test Neutron extensions but not plugins in tree
and that requires renaming the classes
(.*PluginRbacTest => .*ExtRbacTest) to make it clearer.

Having to support all the Neutron plugins in tree is out of scope
for Patrole. Clarifying documentation (which will be done in follow
up) can be added to Patrole documentation to make it clear that
it will only test "main" OpenStack projects, like Tempest.

[0] https://review.openstack.org/#/c/599869/

Change-Id: Iab029f2f875ce2268de12cc2a40e30f2f1a913fe