Fix the branches variant for stable branch job
Keeping explicit mention of stable/pike onwards branches with
all current stable + master end up keep updating that in every cycle.
For example stable/rocky is missed for those job.

This commit use explicit exclusion logic for branches so that we do not
need to update that in every cycle.

Change-Id: I8ff9b5f0648d5f4a145e846315807c730bd20f4b