fix patrole-multinode-admin non-voting gate
The 'patrole-multinode-admin' non-voting gate seems to consistently
fail ServerVolumeAttachmentRbacTest tearDownClass. The failure is coming
from 'test_update_volume_attachment' with the following error message:

Invalid volume: Volume status must be available or error or error_restoring
or error_extending or error_managing and must not be migrating, attached,
belong to a group, have snapshots or be disassociated from snapshots after
volume transfer.'}

The fix is to detach the volume and wait until the detached volume reaches
the 'available' state.

Change-Id: I195115c0d61d15a62cabf3f2b736affbd855cefd