Set admin, member and reader gates to be the same
Since the stein train and usuri and master gates are
running successfully I am proposing to turn on all
of the gates the same and make reader voting.

Add opendev.org to project reference in .zuul.yaml.
to clean up the yaml and make it more like the
tempest .zuul.yaml.

I will do small changes overtime to clean up little
issues and test the gate changes.  Some future area are
tox clean up and looking into the multinode failures
to see if we can make them voting gates.

Removed define concurrency value to allow it to
be set according to the CPU in the VM that the
test are running.

Change-Id: Iad76ad66605a12d55afcc3561e807374ba2943f9