Clean up tox.ini and document zuul.yaml settings
I tried to remove --sitepacakges and the gate failed
so I added comment and in the .zuul.yaml so future people
will understand why.

I Updated the tox.ini environment description so if someone
does a tox -av it will list all command supportted and a
small description of what

While dcoumenting the zuul.yaml I noticed that I made
patrole-reader a job and I feel that we should also
have it as a gate.

I also saw that extension and multi-node non-voting jobs
where missing reader versions so I added them.

Change-Id: I7b8e4c06ece6b32cab4d7238d0bc250334a3adb3