Fix gate: fix the failing network, image, volume RBAC test
1. Image update member test is failing because image owner
itself try to update the image member status but only
admin or that member can update it. Fixing this test.

2. Network tests are failing to create public network
on vxlan or so, fixing those with right parameters.

3. Volume type extra type specs is facing the race condition,
where same name specs is created for all the test which update
delete it. But still it fails so skipping this test too for now.

4. Skipping volume detach test for now, basically squashing
https://review.opendev.org/c/openstack/patrole/+/800594

Story: 2009210
Task: 43272
Story: 2009050
Task: 42820

Change-Id: I5fbcaf219d23d5c94a180c3447ca851d844e1dca